+++ tests/testthat/test-misc.R
@@ -123,16 +123,3 @@ test_that("bind methods work properly", {
   expect_quantities(x, rep(xval, 3), xunt, rep(xval, 3))
   expect_equal(colnames(x), c("a", "a", "a"))
 })
-
-test_that("type_sum is available for quantities objects", {
-  skip_if_not_installed("tibble")
-  library(tibble)
-  expect_equal(type_sum(set_quantities(1, "m", 0.1)), "[(err) m]")
-})
-
-test_that("pillar_shaft is available for quantities objects", {
-  skip_if_not_installed("pillar")
-  library(pillar)
-  expect_equal(as.character(pillar_shaft(set_quantities(1, "m", 0.1))),
-               paste0("1.0", style_subtle("(1)"), " ", style_subtle("m")))
-})